refactor: group AI player support files
