+++ src/components/canvas/players/generation/icons.ts
